Use span for prefetch
