+++ CMakeLists.txt
@@ -14,6 +14,7 @@ set(EXTENSION_SOURCES
     src/cache_prewarm_extension.cpp
     src/core/block_collector.cpp
     src/core/buffer_prewarm_strategy.cpp
+    src/core/os_prefetch.cpp
     src/core/prefetch_prewarm_strategy.cpp
     src/core/prewarm_strategy.cpp
     src/core/prewarm_strategy_factory.cpp
